Fix type mismatch
Backported from NetXMS project.

+++ src/jid.c
@@ -33,7 +33,7 @@ char *xmpp_jid_new(xmpp_ctx_t *ctx, const char *node,
 				    const char *resource)
 {
     char *result;
-    int len,nlen,dlen,rlen;
+    size_t len, nlen, dlen, rlen;
 
     /* jid must at least have a domain */
     if (domain == NULL) return NULL;
@@ -157,7 +157,7 @@ char *xmpp_jid_resource(xmpp_ctx_t *ctx, const char *jid)
 {
     char *result = NULL;
     const char *c;
-    int len;
+    size_t len;
 
     c = strchr(jid, '/');
     if (c != NULL)  {
